Commit to merge

+++ basis-code/no/uio/ifi/asp/parser/AspFloatLiteral.java
@@ -38,6 +38,6 @@ class AspFloatLiteral extends AspAtom{
 
     @Override
     public RuntimeValue eval(RuntimeScope curScope) throws RuntimeReturnValue {
-        return new RuntimeFloatValue(Float.parseFloat(s));
+        return new RuntimeFloatValue(Double.parseDouble(s));
     }
 }
